Move ServiceStack tests to BasicInstrumentation
