+++ tests/Agent/IntegrationTests/IntegrationTests/BasicInstrumentation/ServiceStackApplicationTests.cs
@@ -9,7 +9,7 @@ using NewRelic.Agent.IntegrationTests.RemoteServiceFixtures;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.BasicInstrumentation
 {
     [NetFrameworkTest]
     public class ServiceStackApplicationTests : IClassFixture<ServiceStackApplicationFixture>
