+++ tests/Agent/IntegrationTests/IntegrationTests/BasicInstrumentation/ServiceStackRedisTests.cs
@@ -9,7 +9,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.BasicInstrumentation
 {
     [NetFrameworkTest]
     public class ServiceStackRedisTests : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
